// src/contants/types/items.types.ts
export interface IItemType {
  id: string;
  vpsId: string;
  nicheId: string;
  img: string;
  title: string;
  importLink: string[];
  importPrice: string[];
  ebayPrice: number;
  ebayQuantity: number;
  ebayLink: string;
  ebayId: string;
  note: string;
  updatePriceWM: number;
  timeUpdatePriceWM: string;
  stockWM: string;
  priceOffer: number;
  userId: string;
  createdAt: string;
  updatedAt: string;
  teamId: string;
}